Improve overlay styles
