+++ packages/overlay/src/styles.css
@@ -1,10 +1,21 @@
 .overlay__container {
   position: fixed;
-  inset: 100% 0 0 0;
+  inset: 50% 0 0 0;
+  transform: translateY(100%);
   z-index: 999999;
+  padding: 0 0.5rem 0.5rem 0.5rem;
 }
 .overlay__container[data-open='true'] {
-  inset: 50% 0 0 0;
+  transition: transform 200ms;
+  transform: translateY(0);
+}
+
+.overlay__container__inner {
+  width: 100%;
+  height: 100%;
+  border-radius: 0.5rem;
+  overflow: hidden;
+  /* border: 1px solid #89b6ff; */
 }
 
 .overlay__toggle-button {
