+++ playgrounds/sandbox/vite.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
       // jsxLocation: true,
       name: true,
     }),
-    solidPlugin({}),
+    solidPlugin({ hot: false }),
   ],
   resolve: {
     alias: {
